requirements.txt
fastapi
uvicorn[standard]
pydantic
torch
openai>=1.0.0
python-dotenv
gradio
datasets